requirements.txt
# Core Dependencies
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning & Computer Vision
torchvision>=0.15.0
pillow>=10.0.0
opencv-python>=4.8.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0

# Visualization & Metrics
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Data Processing
h5py>=3.9.0
tqdm>=4.65.0
pyyaml>=6.0

# Evaluation Metrics
scikit-image>=0.21.0
albumentations>=1.3.0

# Utilities
wandb>=0.15.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
